made sure that something was deleted in delete combo, finished first iteration of delete/add custom trick

+++ src/main/java/tkwan/db/ComboDAO.java
@@ -1,6 +1,7 @@
 package tkwan.db;
 
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 public class ComboDAO {
 	private java.sql.Connection conn;
@@ -25,15 +26,34 @@ public class ComboDAO {
 		}
 	}
 	
-	public void deleteCombo(int comboFrom, int comboInto) throws Exception{
+	public boolean deleteCombo(int comboFrom, int comboInto) throws Exception{
 		try {
-			PreparedStatement ps = conn.prepareStatement("DELETE FROM " + tableName + " WHERE comboFrom=? AND comboInto=?;");
+			PreparedStatement ps = conn.prepareStatement("SELECT * FROM " + tableName + " WHERE comboFrom=? AND comboInto=?;");
+			ps.setInt(1, comboFrom);
+			ps.setInt(2, comboInto);
+			ResultSet resultSet = ps.executeQuery();
+			int count = 0;
+			while (resultSet.next()) {
+				count++;
+            }
+			resultSet.close();
+			
+			//nothing to delete
+			if (count == 0) {
+				return false;
+			}
+			
+			ps = conn.prepareStatement("DELETE FROM " + tableName + " WHERE comboFrom=? AND comboInto=?;");
 	        ps.setInt(1, comboFrom);
 			ps.setInt(2, comboInto);
 	        ps.execute();  
             
+	        ps.close();
+	        
+	        return true;
+	        
 		} catch (Exception e) {
-            throw new Exception("Failed in adding combo: " + e.getMessage());
+            throw new Exception("Failed in deleting combo: " + e.getMessage());
 		}
 	}
 }
